requirements/base.txt
alabaster==0.7.7
amqp==1.4.9
anyjson==0.3.3
arrow==0.7.0
Babel==2.3.4
backports.shutil-get-terminal-size==1.0.0
billiard==3.3.0.23
binaryornot==0.4.0
celery==3.1.21
chardet==2.3.0
click==6.6
colorama==0.3.7
cookiecutter==1.4.0
coverage==4.0.3
decorator==4.0.9
defusedxml==0.4.1
Django==1.9.4
django-allauth==0.24.1
django-autoslug==1.9.3
django-braces==1.8.1
django-coverage-plugin==1.2.2
django-crispy-forms==1.6.0
django-debug-toolbar==1.4
django-environ==0.4.0
django-extensions==1.6.1
django-filter==0.13.0
django-floppyforms==1.6.1
django-model-utils==2.4
django-pipeline==1.6.8
django-pipeline-browserify==0.4.0
django-redis==4.3.0
django-secure==1.0.1
django-test-plus==1.0.12
djangorestframework==3.3.3
docutils==0.12
factory-boy==2.6.1
fake-factory==0.5.7
future==0.15.2
imagesize==0.7.1
ipdb==0.9.0
ipython==4.2.0
ipython-genutils==0.1.0
Jinja2==2.8
jinja2-time==0.1.0
kombu==3.0.35
Markdown==2.6.6
MarkupSafe==0.23
oauthlib==1.0.3
pickleshare==0.7.2
Pillow==3.2.0
poyo==0.2.0
psycopg2==2.6.1
py==1.4.31
Pygments==2.1.3
pytest==2.9.1
pytest-django==2.9.1
pytest-sugar==0.5.1
python-dateutil==2.5.3
python3-openid==3.0.10
pytz==2015.7
redis==2.10.5
requests==2.9.1
requests-oauthlib==0.6.1
simplegeneric==0.8.1
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.4.1
sqlparse==0.1.19
termcolor==1.1.0
traitlets==4.2.1
unicode-slugify==0.1.3
Unidecode==0.4.19
Werkzeug==0.11.4
wheel==0.29.0
whichcraft==0.3.1
whitenoise==2.0.6
